+++ src/core/se/jbee/inject/service/ServiceMalfunction.java
@@ -1,0 +1,9 @@
+package se.jbee.inject.service;
+
+public class ServiceMalfunction extends RuntimeException {
+
+	public ServiceMalfunction(String message, Throwable cause) {
+		super(message, cause);
+	}
+	
+}